# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api to manage a fast food restaurant",
        "title": "Tc-Fiap-50",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/customer": {
            "get": {
                "description": "Get customer by CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Get customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CPF",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetCustomerResponseDto"
                        }
                    }
                }
            },
            "post": {
                "description": "Add customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Add customer",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddCustomerRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddCustomerRequestDto": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "integer",
                    "example": 12345678901
                },
                "email": {
                    "type": "string",
                    "example": "john@doe.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "dto.GetCustomerResponseDto": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}